changed: morre succint enum implementation for the ISR definitions. fixed: ioPtr int repr should be u16.

+++ src/avr_io/private/mapped_io.nim
@@ -2,13 +2,13 @@ import volatile
 
 
 type MappedIoRegister8* = distinct ptr uint8
-type MappedIoRegister16* = distinct ptr uint
+type MappedIoRegister16* = distinct ptr uint16
 
 template `[]`*(p: MappedIoRegister8): uint8 =
   volatile.volatileLoad((ptr uint8)p)
 
 template `[]=`*(p: MappedIoRegister8, v: uint8) =
-  volatile.volatileStore((ptr uint16)p, v)
+  volatile.volatileStore((ptr uint8)p, v)
 
 template `[]`*(p: MappedIoRegister16): uint16 =
   volatile.volatileLoad((ptr uint16)p)
@@ -16,7 +16,7 @@ template `[]`*(p: MappedIoRegister16): uint16 =
 template `[]=`*(p: MappedIoRegister16, v: uint16) =
   volatile.volatileStore((ptr uint16)p, v)
 
-template ioPtr8(a: uint8): ptr uint8 = 
+template ioPtr8(a: uint16): ptr uint8 = 
   cast[ptr uint8](a)
 
 template ioPtr16(a: uint16): ptr uint16 =
